Update TCA9548A.cpp
Further ambiguity discrepancies between TwoWire library MCU iterations. 

This allows the Mega to compile, while also the Uno and ESP32 (tested).

+++ src/TCA9548A.cpp
@@ -74,7 +74,7 @@ void TCA9548A::write(uint8_t inData)
 uint8_t TCA9548A::read()
 {
     uint8_t buff = 0;
-    this->myWire->requestFrom((uint16_t)this->_address, (uint8_t)1, true);
+    this->myWire->requestFrom((uint16_t)this->_address, (uint8_t)1, (uint8_t)true);
 
     if(!this->myWire->available())
         return 255;
